+++ src/test/java/org/realityforge/replicant/server/transport/PacketQueueTest.java
@@ -42,6 +42,43 @@ public class PacketQueueTest
     assertEquals( queue.getPacket( 4 ).getSequence(), 4 );
   }
 
+  @Test
+  public void ack()
+  {
+    final PacketQueue queue = new PacketQueue();
+    assertEquals( queue.getLastSequenceAcked(), 0 );
+    assertEquals( queue.size(), 0 );
+
+    queue.addPacket( null, null, newChanges() );
+    queue.addPacket( null, null, newChanges() );
+    queue.addPacket( null, null, newChanges() );
+    queue.addPacket( null, null, newChanges() );
+
+    assertEquals( queue.size(), 4 );
+
+    queue.ack( 0 );
+    assertEquals( queue.getLastSequenceAcked(), 0 );
+    queue.ack( 1 );
+    assertEquals( queue.getLastSequenceAcked(), 1 );
+    assertEquals( queue.size(), 3 );
+    queue.ack( 0 );
+    assertEquals( queue.getLastSequenceAcked(), 1 );
+    assertEquals( queue.size(), 3 );
+    queue.ack( 4 );
+    assertEquals( queue.getLastSequenceAcked(), 4 );
+    assertEquals( queue.size(), 0 );
+
+    try
+    {
+      queue.ack( 5 );
+      fail( "Attempted to ack future sequence" );
+    }
+    catch ( final IllegalStateException ise )
+    {
+      //Ignored
+    }
+  }
+
   private ArrayList<EntityMessage> newChanges()
   {
     return new ArrayList<>();